update: 30.05.2022

+++ crop_layout/src/main/java/ua/com/compose/image_maker/data/AspectRatio.kt
@@ -3,7 +3,7 @@ package ua.com.compose.image_maker.data
 sealed class Ratio {
     object Custom : Ratio()
     object OriginRatio : Ratio()
-    data class AspectRatio(val first: Int, val second: Int): Ratio()
+    data class AspectRatio(val first: Float, val second: Float): Ratio()
 }
